Fix footer layout regression
Commit a974e082b9a884e0119939391aafc6f7f9038f59 caused a regression
where the footer would not correctly be rendered as two boxes stacked on
each other on small browser sizes. The footer content was always
rendered in horizontally stacked boxes which looked like crap when there
was not much horizontal space. See this comment for a screenshot:

https://github.com/jointwt/twtxt/pull/321#issuecomment-752348483

The original behavior was flawed, too. Only very small browser sizes
would work. Now we render it only horizontally, if the whole contents
fits in a single line. Whenever we need more than one, we switch to
vertical stacking. This obviously depends on the chosen font available
on the system, so my selected width of 1200px may be a bit more
conservative than necessary.